loop range fix and refactor
