+++ testdata/src/go.uber.org/slices/slices.go
@@ -365,87 +365,6 @@ func testDoubleLenCheck(a, b []int) int {
 	return 0
 }
 
-// this test makes sure that all forms of range loops correctly produce their ranging expression
-// as non-nil - including limiting the scope of that production to within their loop bodies
-
-func testRangeProducesNonnil(a []*int) *int {
-	for range a {
-		// here and in following similar cases:
-		// check that a can be indexed into (since it's nonnil)
-		// validity of the index is outside the scope of NilAway, so we can use any index, e.g. 0
-		return a[0]
-	}
-	if dummyBool() {
-		return a[0] //want "sliced into"
-	}
-	for _ = range a {
-		return a[0]
-	}
-	if dummyBool() {
-		return a[0] //want "sliced into"
-	}
-	for i := range a {
-		dummyConsume(i)
-		return a[0]
-	}
-	if dummyBool() {
-		return a[0] //want "sliced into"
-	}
-	for _, _ = range a {
-		return a[0]
-	}
-	if dummyBool() {
-		return a[0] //want "sliced into"
-	}
-	for i, _ := range a {
-		dummyConsume(i)
-		return a[0]
-	}
-	if dummyBool() {
-		return a[0] //want "sliced into"
-	}
-	for _, j := range a {
-		dummyConsume(j)
-		return a[0]
-	}
-	if dummyBool() {
-		return a[0] //want "sliced into"
-	}
-	for i, j := range a {
-		dummyConsume(i)
-		dummyConsume(j)
-		return a[0]
-	}
-	if dummyBool() {
-		return a[0] //want "sliced into"
-	}
-	var i2, j2 interface{}
-	for i2, _ = range a {
-		dummyConsume(i2)
-		return a[0]
-	}
-	if dummyBool() {
-		return a[0] //want "sliced into"
-	}
-	for _, j2 = range a {
-		dummyConsume(j2)
-		return a[0]
-	}
-	if dummyBool() {
-		return a[0] //want "sliced into"
-	}
-	for i2, j2 = range a {
-		dummyConsume(i2)
-		dummyConsume(j2)
-		return a[0]
-	}
-	if dummyBool() {
-		return a[0] //want "sliced into"
-	}
-	i := 0
-	return &i
-}
-
 func testSwitchAsLenCheck(a []int) int {
 	var i int
 	switch len(a) {
